\name{make.col.vec}
\alias{make.col.vec}

\title{
  make color vector
}
\description{
 utility function to make a vector of colors used for plotting
}
\usage{
make.col.vec(x, colors)
}

\arguments{
  \item{x}{
    vector of 0, 1, 2's

}
\item{colors}{
  vector of "col1", "col2" , "col3"

}
}

\value{
 returns vector of colors 
}
