\name{IBDhaploRtools-package}
\alias{IBDhaploRtools-package}
\alias{IBDhaploRtools}
\docType{package}
\title{
Functions for the Analysis of IBD Haplo output
}
\description{
IBDhaploRtools consists of several functions to analyze, plot, and store the output of the IBD_Haplo software package. More information regarding IBD_Haplo can be found at www.stat.washington.edu/thompson/Genepi/pangaea.shtml.
}
\details{
\tabular{ll}{
Package: \tab IBDhaploRtools\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2014-06-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

See the package vignette: vignette( "IBDhaploRtools_tutorial" ) 

}
\author{
  Marshall D. Brown

  Fiona Grimson
  <fgrimson@uw.edu>

}
\references{
  none
}


\examples{
# See the tutorial that  outlines the use of all the
# functions found in this package: 
# vignette( IBDhaploRtools_tutorial )
}
